pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "cumpyl"
version = "0.3.0"
description = "A binary rewriting tool with encoding/decoding capabilities"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Cumpyl Framework"}
]
maintainers = [
    {name = "Cumpyl Framework"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers", 
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]
requires-python = ">=3.9"
dependencies = [
    "lief",
    "capstone", 
    "keystone-engine",
    "rich",
    "tenacity",
    "tqdm",
    "bitsandbytes",
    "cryptography",
    "pyyaml",
    "textual",
    "jsonschema"
]

[project.optional-dependencies]
dev = [
    "numpy",
    "pytest",
    "pytest-cov", 
    "black",
    "flake8"
]
test = [
    "numpy"
]

[project.scripts]
cumpyl = "cumpyl_package.cumpyl:main"

[project.urls]
Homepage = "https://github.com/umpolungfish/cumpyl"
Repository = "https://github.com/umpolungfish/cumpyl.git"
Issues = "https://github.com/umpolungfish/cumpyl/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["cumpyl_package*"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=cumpyl_package"
